Move IPhysicalGpu to separate header file to hide NVAPI specific library data types.
