+++ NVAPIHooks/Header Files/PhysicalGpu.h
@@ -3,24 +3,11 @@
 #pragma once
 
 #include <Export.h>
+#include <IPhysicalGpu.h>
 #include <nvapi.h>
 
 namespace NVAPIHooks
 {
-	class IPhysicalGpu
-	{
-	public:
-		IPhysicalGpu() = default;
-		~IPhysicalGpu() = default;
-		IPhysicalGpu(IPhysicalGpu&) = delete;
-		IPhysicalGpu(IPhysicalGpu&&) = delete;
-		IPhysicalGpu& operator = (IPhysicalGpu&) = delete;
-		IPhysicalGpu& operator = (IPhysicalGpu&&) = delete;
-
-		/// <returns>Number of cores defined for the GPU.</returns>
-		virtual unsigned long GetCoreCount() = 0;
-	};
-
 	class PhysicalGpu : public IPhysicalGpu
 	{
 	public:
