+++ NVAPIHooks/NVAPIHooks.vcxproj.filters
@@ -29,6 +29,9 @@
     <ClInclude Include="Header Files\PhysicalGpu.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="Header Files\IPhysicalGpu.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Source Files\pch.cpp">
